Allow unregistered names, CLI now runs blocks
The old name system had a problem that made it awkward to send names and
definitions in the same block. For example, consider this block:

    name foo
    bond foo() : #word { #0 }

When serializing this block, the user's client would run the block
locally. This run would assign an index to "foo" (for example, #47), and
then, for each occurrence of "foo" afterward, it would be serialized
using the index (here, #47). But what if some other block is mined
before this user's block? In that case, other names would be registered
*before* foo. As such, other name would get the index #47, and the
serialization of `bond foo()` would be deserialized incorrectly.

The only option would be to first submit the names and, once they are
mined, submit the type or bond you want to deploy. But this would take a
lot of time. To simplify this, names are, by default, serialized in full
when submitted by users. Miners can still use the "name" feature to
compress blocks that they receive.

A similar issue happens when an user attempts to serialize a "case-of"
expression. Since the serialization depends on the datatype, if the
datatype isn't defined, then the serialization is meaningless. An user
may attempt to submit both the type and the bond that uses it at the
same time, but nothing guarantees the type will be mined before the
bond (or, worse, that another type with the same name will be mined). As
such, users must send bonds after the types they use are mined.

Since the use-case of sending types and bonds that use these types in
the same block isn't available, the parser will no longer attempt to
define datatypes when parsing a block. That means that, for example,
trying to parse this in the same block won't work:

    type Nat {
      zero{}
      succ{pred: Nat}
    }

    bond double(x: Nat): Nat {
      case x : Nat {
        zero: zero{}
        succ: succ{succ{double(x.pred)}}
      }
    }

Because, when attempting to parse the `case x : Nat` expression, the
parser won't have the `Nat` definition saved, and won't know how to
parse it. This will simplify the parser, since defining types at
parse-time was a weird aspect it had before. The snippet above can still
work, as long as it is split in two blocks:

    type Nat {
      zero{}
      succ{pred: Nat}
    }

    save

    bond double(x: Nat): Nat {
      case x : Nat {
        zero: zero{}
        succ: succ{succ{double(x.pred)}}
      }
    }

    save

Note that the "save" keyword is not a native operation. It is just a
delimiter used by the debugging tool to separate blocks.