Mutable files instead of bonds
Kelvin observed that rebinding global definitions interacts very poorly
with lazy evaluation. For example,

    bond inc_count(): & #word {
      bind get_count {
        +(get_count(), #1)
      }
      return get_count()
    }

    eval {
      run a : #word : inc_count()
      run b : #word : inc_count()
      return a
    } : & #word

Here, the evaluation will return "2", even though it looks like it
should return "1". That's because `inc_count()` will lazily evaluate to
`get_count()`, so, `return a` is actually `return get_count()`, which
would resolve after both increments happened. I agree this problem would
be a major source of confusion and bugs, so, in general, we should
absolutely avoid mutating global constants and bonds.

As such, I've added a new kind of entry, "file", for mutable state. The
key difference is that a file can't be called inside pure terms (i.e.,
it doesn't interact with the "call" constructor). The only way to
retrieve a file's contents is by using the new "get" effect. Since that
effect is part of the Effect monad, the order of evaluation will be as
expected.

The README contains examples on how to declare, retrieve and mutate
files.

This commit also disables name compression. I'll revisit this later.